+++ web-wallet/src/lib/errors/index.js
@@ -1,0 +1,3 @@
+export { default as InvalidMnemonicError } from "./invalid-mnemonic-error";
+export { default as InvalidPasswordError } from "./invalid-password-error";
+export { default as MismatchedWalletError } from "./mismatched-wallet-error";