+++ web-wallet/src/lib/errors/invalid-password-error.js
@@ -1,0 +1,8 @@
+class InvalidPasswordError extends Error {
+  constructor() {
+    super("Wrong password");
+    this.name = "InvalidPasswordError";
+  }
+}
+
+export default InvalidPasswordError;